Added logout, usePersistedState; refactor AuthContext

+++ client/src/components/itemDetails/ItemDetails.jsx
@@ -1,12 +1,14 @@
 import { Link, useParams } from "react-router-dom"
 
 import { useDeleteItem, useGetSingleItem } from "../../hooks/useItems";
+import { useAuthContext } from "../../contexts/AuthContext";
 
 import "./itemDetails.css"
 
 export default function ItemDetails() {
     const { itemId } = useParams();
     const [item] = useGetSingleItem(itemId);
+    const { isOwner } = useAuthContext();
 
     const deleteItem = useDeleteItem();
 
@@ -18,8 +20,8 @@ export default function ItemDetails() {
             <p>{item.price}$</p>
             <p>{item.description}</p>
             <div>
-                <button onClick={async () => await deleteItem(itemId)} className="button-secondary">Delete</button>
-                <Link to={`/edit/${itemId}`} className="button-main">Edit</Link>
+                {isOwner(itemId) && <><button onClick={async () => await deleteItem(itemId)} className="button-secondary">Delete</button>
+                    <Link to={`/edit/${itemId}`} className="button-main">Edit</Link></>}
             </div>
         </>
     )
